requirements.txt
alembic==1.11.1
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
feedparser==6.0.10
greenlet==2.0.2
idna==3.4
markdownify==0.11.6
PyJWT==2.7.0
python-dotenv==1.0.0
pythorhead==0.10.0
pyyaml==6.0.0
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==2.0.3
